feat: provide default libp2p instance
The original intention was to allow users to configure a libp2p node
to their requirements but it's a non-trivial undertaking unless the
user is deeply familiar with the libp2p stack.

Instead, let's provide node and browser libp2p instances with sensible
defaults that give the user the best chance of success on first try.