interop: add webtransport address to Kubo
We (the Kubo maintainers) to do not promise that anything will work with browsers without webtransport.
I get that it happen to work now since the websocket is over localhost but still let's include webtransport since we at least want to check having it on does not make anything worst.