chore: pull block to pin from relay server instead of peer
We previously had a circuit relay connection to the peer in the
test, but circuit relay connections are transient (e.g. time/data
limited) so we can't run bitswap over it.

Instead pull the block from the relay server as we can have a regular
(non-transient) websocket connection to that peer.