fix: try circuit relay transport first
Make circuit relay the first transport to try so it will match addresses
ahead of supporting transports.

E.g. where the address is `/<webtransport>/<circuit-relay>` get the
circuit relay transport to dial first as it will strip off the relay
part, hand control back to the transport manager and then run the
circuit relay protocol over the newly created webtransport connection.

This is a tactical fix for a bug that needs to be addressed in libp2p.