feat!: use helia router for IPNS put/get
Uses the configured `.routing` property on the Helia interface to
put/get IPNS records by default.

Removes the `libp2p` routing as `helia` configures this as a
`.routing` implementation and `@helia/http` doesn't have it.

Updates the description of the `pubsub` routing to make it's
limitations clearer.

BREAKING CHANGE: `helia.routing` is the default routing used, the `libp2p` routing has been removed as it is redundant