fix: remove stubs from verified-fetch tests
The extensive stubbing in the `@helia/verified-fetch` tests have some
baked-in assumptions about how the codecs work which are not easy
to unpick.

It's quick to test using the actual codecs if the block data is already
present so remove the stubs and use a network-less Helia node to make
the tests more reliable.