chore: fix multiblock car file test
The test introduces an artificially small block size to ensure we
have a CAR file with lots of blocks but it's quite slow and often
causes CI failures.

The large file being added to the CAR file is already big enough
to be formed of multiple blocks so splitting it into many smaller
blocks doesn't give us much.