fix: Add GlobSourceResult to globSource return type in unixfs. (#475)
Modifies the return type on the `globSource` utility in `@helia/unixfs` to return the `GlobSourceResult` in addition to the `ImportCandidate` type. This means that the `path` property will be a `string` instead of `string | undefined`.

Makes it a bit easier to use `globSource` since it is a bit annoying to have to check if `path` is defined when it always is.