fix: use a short-lived AbortSignal for fetch operations
Node's `undici` module appears to be adding listeners for the `"abort"` event on any passed AbortSignal but then not cleaning them up.

This casues many `MaxListenersExceededWarning`s to appear in the console and likely causes a memory leak.

To work around this, use a short-lived `AbortSignal` for each request that can be garbage collected along with the request itself.

See https://github.com/nodejs/node/issues/52635 for more information.