fix: support wrapWithDirectory in addFile
If the user passes a `FileCandidate` with a `path` property or passes
the `wrapWithDirectoryOption` the return type of `fs.addFile` will
be a `DirectoryEntry`.

This breaks existing behaviour where it will always be a `FileEntry`

Since `URLSource` returns a `{ content, path }` candidatem this can
cause a `DirectoryEntry` to be returned whereas it used to always be
a `FileEntry`.  An `ignorePath` option has been added to restore the
previous behavior.

Fixes #707

BREAKING CHANGE: addFile can sometimes return a directory, URLSource returns directory when a path is present