fix!: require path in addFile call (#754)
To ensure metadata is always stored, the CID returned from `addFile` will resolve to a directory.

The `path` property is now required of the import candidate passed to `addFile`, if it's not necessary to store a path as part of a file, use `addBytes` or `addByteStream`.

Since `urlSource` returns a `{ content, path }` candidate, it can be used with `addFile`.

To import directly from a URL without a path, a `urlByteStream` function has been added for use with `addByteStream`.

Fixes #707

BREAKING CHANGE: addFile requires the import candidate to have `path` and `content` properties and the returned CID will always resolve to a directory

---------

Co-authored-by: Daniel N <2color@users.noreply.github.com>